refactor: standardize test file extensions and improve type safety
- Migrate all test files from .spec.ts to .test.ts extensions for consistency
- Replace any types with proper JsonValue | null types in App.tsx
- Clean up unused variables in stdinHandler.ts with proper underscore prefixes
- Remove debug console.log statements from appService.ts and stdinHandler.ts
- Consolidate syntax highlighting tests into single comprehensive test file
- All tests passing with improved type safety and code quality
